Delete all boilerplate from CRA
